requirements.txt
Django==5.0.7
djangorestframework==3.15.2
Pillow==10.4.0
